Remove references to YGtPM() class and objects
This class was simply used to group functions together. This is not
ideal because classes should really be reserved for creating objects
with attributes and methods. A better way of grouping related functions
together in more functional programming is to split the functions up
into modules within a package (see
https://stackoverflow.com/questions/38758668/grouping-functions-by-using-classes-in-python).
Another possibility is to use the @staticmethod declaration for methods
in the class, but I have seen some references stating that this is
rarely needed and should be used sparingly (see
https://stackoverflow.com/questions/735975/static-methods-in-python).

Consequently, references to the YGtPM() class and objects were removed
from ymap.py and tests.py. Hopefully, it will be possible to organize
code into modules of related functions at a later stage.