Refactor pmap()
Having factored out the code that processed the yeastID file into the
parse_gene_names() function, we now remove the the redundant code within
pmap().

Updated docstring (conform to PEP257) to describe the function
parameters. We specify, as a reminder, that the yeastID_input parameter
will now be a dictionary (or suitable map-like object).