Refactor ptm_map()
Having factored out the code that processed the yeastID file into the
parse_mutations() function, we now remove the the redundant code within
ptm_map().

Updated docstring (conform to PEP257) to describe the function
parameters. We specify, as a reminder, that the mut_prot_inptu parameter
will now be a dictionary (or suitable map-like object).