Implement dmap()
Rewrite docstring (including parameter description). Conform to PEP257.
Give variables sensible, informative names. Eliminate redundancy. More
Pythonic.

Loop over mutation positions for every mutated protein in mut_prot_input
dictionary. Check whether each mutation position is in a domain of the
mutated protein. If it is, store the mutation data and data about the
domain to which the mutation mapped as strings. Put these strings in a
list of lines to be written to the output_file and then write them once
at the very end.