Implement enrich()
Rewrite docstring. Conform to PEP257. Give variables sensible,
informative names.

Keep file with mapped mutations open for a little as possible. Write
lines to file once at the end. Note that the call
'annotations.get_enriched_terms(sgd_gene_names)' is the bottleneck to
speed here. Testing suggests it takes approximately 20s to run. The
ontology and annotation objects are instantiated outside the function
once, so that they can be reused if enrich() is called on different
files. It may be important to include an option in commands that lets
users specify whether they want to perform GO enrichment.