Implement mmap()
Rewrite docstring (including parameter description). Conform to PEP257.
Give variables sensible, informative names. Eliminate redundancy. More
Pythonic.

Loop over mutation positions for every mutated protein in mut_prot_input
dictionary. Check if each mutation position lands affects an important
residue in an binding site or active site of the mutated protein. If it
is, generate a list of mutation data and data about the active/binding
site to which the mutation mapped. Convert this list to a line (string)
and append the line to a list of lines to be written to the output_file
and then write them once at the very end.