Implement nucleotide_map()
Rewrite docstring (including parameter description). Conform to PEP257.
Give variables sensible, informative names. Eliminate redundancy. More
Pythonic.

Loop over mutation positions for every mutated protein in mut_prot_input
dictionary. Check if each mutation position lands in a nucleotide
binding site. If it does, generate a list of mutation data and data
about the nucleotide binding site to which the mutation mapped. Convert
this list to a line (string). Append the line to a list of lines that
will be written to the output_file once at the very end.