Implement pdb()
Map the positions of mutations in mutated proteins to structural
elements (helices, beta strands, turns).

Rewrite docstring (including parameter description). Conform to PEP257.
Give variables sensible, informative names. Eliminate redundancy. More
Pythonic.

Loop over mutation positions for every mutated protein in mut_prot_input
dictionary. Check if each mutation position lands in a structural
element of the mutated protein. If it does, generate a list of mutation
data and data about the structural element to which the mutation mapped.
Convert this list to a line (string) and append the line to a list of
lines to be written to the output_file once at the very end.