Implement parse_names_by_locus()
This function takes the gene_names dictionary produced by
parse_gene_names() and parses it into a new dictionary whose keys are
locus names and values are lists [uniprot_id, locus_name, ]. The
redundancy of storing the key as an element of the value list is used to
facilitate the production of a line (that will be written to a file).

This allows us to refactor code out of id_map(), interface() and ppi().