Fix error in sc_btw_proteins.txt input file and recompress into zip
During testing, found a KeyError for lookup of locus name/id 'YMR5C'. Upon insepection, it is clear that the gene name BUL1 corresponds to locus name/id 'YMR275C' (see SGD website). I  changed it manually, but need to check whether this is present in up-to-date downloaded sc_between_proteins.txt file (could cause future problems). If it is found to be a problem with the PTMcode database (from which the file originates), contact the maintainers of the database.