Fix bugs in enrich()
Bug 1 - Where line could be referenced before it was assigned to. Need to append
lines within for loop, to capture all the significantly enriched GO
terms.

Bug 2 - Check whether, after looping over all the enriched GO terms, any lines have been added to the lines list. If they haven't this means that none of the GO terms are SIGNIFICANTLY enriched. So we write a message in the results file to say this.